moved config into its own package
