+++ src/main/java/jamdoggie/betterbattletowers/BattleTowerMod.java
@@ -1,6 +1,7 @@
 package jamdoggie.betterbattletowers;
 
 import jamdoggie.betterbattletowers.block.BattleTowerBlocks;
+import jamdoggie.betterbattletowers.config.BattleTowerConfig;
 import jamdoggie.betterbattletowers.entity.MobAgressiveZombiePig;
 import jamdoggie.betterbattletowers.entity.MobGolem;
 import jamdoggie.betterbattletowers.block.TileEntityChestTower;
@@ -10,6 +11,7 @@ import jamdoggie.betterbattletowers.worldgen.util.TowerProperties;
 import jamdoggie.betterbattletowers.worldgen.WorldFeatureBattleTower;
 import net.fabricmc.api.ClientModInitializer;
 import net.fabricmc.api.ModInitializer;
+import net.minecraft.client.gui.guidebook.mobs.MobInfoRegistry;
 import net.minecraft.client.sound.SoundRepository;
 import net.minecraft.core.item.ItemStack;
 import net.minecraft.core.item.Items;
@@ -24,6 +26,7 @@ import turniplabs.halplibe.util.*;
 
 import java.util.Random;
 
+import static jamdoggie.betterbattletowers.worldgen.util.LootTable.LAPIZ;
 import static net.minecraft.core.net.command.util.CommandHelper.registerWorldFeatureClass;
 
 
@@ -56,6 +59,17 @@ public class BattleTowerMod implements ModInitializer, GameStartEntrypoint, Reci
 
 	@Override
 	public void afterClientStart() {
+		ItemStack lapiz = new ItemStack(Items.DYE);
+		lapiz.setMetadata(LAPIZ);
+		MobInfoRegistry.register(MobGolem.class, "betterbattletowers.golem.name", "betterbattletowers.golem.desc", MobGolem.MAX_HEALTH, 10000,
+			new MobInfoRegistry.MobDrop[]{
+				new MobInfoRegistry.MobDrop(new ItemStack(Items.DIAMOND), 1.0f, 1, 5),
+				new MobInfoRegistry.MobDrop(new ItemStack(Items.ORE_RAW_IRON), 1.0f, 0, 3),
+				new MobInfoRegistry.MobDrop(new ItemStack(Items.DUST_REDSTONE), 1.0f, 0, 5),
+				new MobInfoRegistry.MobDrop(new ItemStack(Items.OLIVINE), 1.0f, 0, 3),
+				new MobInfoRegistry.MobDrop(lapiz, 1.0f, 0, 5),
+				new MobInfoRegistry.MobDrop(new ItemStack(Items.ORE_RAW_GOLD), 1.0f, 0, 3)
+			});
 	}
 
 	@Override
